+++ src/filter/map/index.ts
@@ -7,18 +7,16 @@ import type { MatchMode } from "../../types";
  * @param i - The index to be used in the pattern string.
  * @param matchMode - The mode of matching to be applied.
  * @returns A string representing the SQL pattern based on the provided match mode.
+ * @example
+ * // Returns "$1,$2"
+ * index([1, 2], undefined);
+ * @example
+ * // Returns "($1,$2)"
+ * index([1, 2], "in");
  */
 function index(index: number[], matchMode: MatchMode | undefined): string {
-  const i = index.map((i: number) => `$${i}`) ;
+  const i = index.map((i: number) => `$${i}`);
   switch (matchMode) {
-    case "startsWith":
-      return `${i}%`;
-    case "endsWith":
-      return `%${i}`;
-    case "contains":
-      return `%${i}%`;
-    case "notContains":
-      return `%${i}%`;
     case "in":
       return `(${i})`;
     default: